replication/qualsyn-1.0.0/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qualsyn"
version = "1.0.0"
description = "Quality-Driven Synthetic Data Generation via LLM-Guided Oversampling"
readme = "README.md"
license = {text = "GPL-3.0"}
authors = [
    {name = "Asim Sinan Yuksel", email = "asimyuksel@sdu.edu.tr"}
]
keywords = [
    "synthetic data",
    "oversampling",
    "imbalanced learning",
    "large language models",
    "quality validation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "imbalanced-learn>=0.11.0",
    "xgboost>=1.7.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/asimyuksel/qualsyn"
Documentation = "https://qualsyn.readthedocs.io"
Repository = "https://github.com/asimyuksel/qualsyn"

[tool.setuptools.packages.find]
where = ["."]
include = ["qualsyn*"]

